refactor(docker): move docker components to react [EE-3348]
closes [EE-3348]
