--- app/docker/networks/edit/index.ts
@@ -1,5 +1,0 @@
-import { react2angular } from '@/react-tools/react2angular';
-
-import { NetworkDetailsView } from './NetworkDetailsView';
-
-export const NetworkDetailsViewAngular = react2angular(NetworkDetailsView, []);
