--- app/docker/networks/index.ts
@@ -1,7 +1,0 @@
-import angular from 'angular';
-
-import { NetworkDetailsViewAngular } from './edit';
-
-export const networksModule = angular
-  .module('portainer.docker.networks', [])
-  .component('networkDetailsView', NetworkDetailsViewAngular).name;
