+++ app/docker/react/components/index.ts
@@ -1,6 +1,29 @@
 import angular from 'angular';
 
-export const componentsModule = angular.module(
-  'portainer.docker.react.components',
-  []
-).name;
+import { r2a } from '@/react-tools/react2angular';
+import { ContainersDatatableContainer } from '@/react/docker/containers/ListView/ContainersDatatable/ContainersDatatableContainer';
+import { ContainerQuickActions } from '@/react/docker/containers/components/ContainerQuickActions';
+
+export const componentsModule = angular
+  .module('portainer.docker.react.components', [])
+  .component(
+    'containersDatatable',
+    r2a(ContainersDatatableContainer, [
+      'endpoint',
+      'isAddActionVisible',
+      'dataset',
+      'onRefresh',
+      'isHostColumnVisible',
+      'tableKey',
+    ])
+  )
+  .component(
+    'containerQuickActions',
+    r2a(ContainerQuickActions, [
+      'containerId',
+      'nodeName',
+      'state',
+      'status',
+      'taskId',
+    ])
+  ).name;
