+++ app/react/docker/containers/ListView/ContainersDatatable/columns/stack.tsx
@@ -1,6 +1,6 @@
 import { Column } from 'react-table';
 
-import type { DockerContainer } from '@/docker/containers/types';
+import type { DockerContainer } from '@/react/docker/containers/types';
 
 export const stack: Column<DockerContainer> = {
   Header: 'Stack',
