+++ app/react/docker/containers/components/ContainerQuickActions/ContainerQuickActions.tsx
@@ -1,8 +1,8 @@
 import clsx from 'clsx';
 
+import { DockerContainerStatus } from '@/react/docker/containers/types';
 import { Authorized } from '@/portainer/hooks/useUser';
 import { react2angular } from '@/react-tools/react2angular';
-import { DockerContainerStatus } from '@/docker/containers/types';
 
 import { Link } from '@@/Link';
