+++ app/react/docker/networks/ItemView/NetworkContainersTable.tsx
