+++ app/react/docker/networks/ItemView/NetworkDetailsTable.test.tsx
