+++ app/react/docker/networks/ItemView/NetworkDetailsTable.tsx
