+++ app/react/docker/networks/ItemView/index.ts
@@ -1,0 +1,1 @@
+export { ItemView } from './ItemView';
